feat: schedult entity 연관관계 매핑#
