+++ src/main/java/unilearn/unilearn/study/entity/Regist.java
@@ -1,8 +1,11 @@
-package unilearn.unilearn.user.entity;
+package unilearn.unilearn.study.entity;
 import lombok.*;
+import unilearn.unilearn.user.entity.User;
 
 import javax.persistence.*;
 import javax.transaction.Transactional;
+import java.time.LocalDateTime;
+
 
 @Entity
 @Getter
@@ -10,10 +13,11 @@ import javax.transaction.Transactional;
 @NoArgsConstructor(access = AccessLevel.PUBLIC)
 @AllArgsConstructor
 @Transactional
-public class StdList {
+
+public class Regist {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long std_List_id;
+    private Long regist_id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
@@ -28,4 +32,15 @@ public class StdList {
 //    @JoinColumn(name = "subject_id", nullable = false)
 //    private Subject subject;
 
+    @Column(nullable = false)
+    private String regist_detail;
+
+    @Column(nullable = false)
+    private LocalDateTime created_at;
+
+    @Column(nullable = false)
+    private LocalDateTime updated_at;
+
+    @Column(nullable = false)
+    private String regist_status;
 }
