+++ src/main/java/unilearn/unilearn/study/entity/StdList.java
@@ -1,10 +1,9 @@
-package unilearn.unilearn.user.entity;
+package unilearn.unilearn.study.entity;
 import lombok.*;
+import unilearn.unilearn.user.entity.User;
 
 import javax.persistence.*;
 import javax.transaction.Transactional;
-import java.time.LocalDateTime;
-
 
 @Entity
 @Getter
@@ -12,11 +11,10 @@ import java.time.LocalDateTime;
 @NoArgsConstructor(access = AccessLevel.PUBLIC)
 @AllArgsConstructor
 @Transactional
-
-public class Regist {
+public class StdList {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long regist_id;
+    private Long std_List_id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
@@ -27,19 +25,8 @@ public class Regist {
     private Study study;
 
 //    subject 테이블 부재
-//    @ManyToOne(fetch = FetchType.LAZY)
-//    @JoinColumn(name = "subject_id", nullable = false)
-//    private Subject subject;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "subject_id", nullable = false)
+    private Subject subject;
 
-    @Column(nullable = false)
-    private String regist_detail;
-
-    @Column(nullable = false)
-    private LocalDateTime created_at;
-
-    @Column(nullable = false)
-    private LocalDateTime updated_at;
-
-    @Column(nullable = false)
-    private String regist_status;
 }
